Mapped models to views
Created EntitySprites that hold GridEntitys and display them in world space.
Created a playersprite object and animated it.
Created background and matched the grid to world space.
Entities now know who they are facing and if two entites are facing each other.
Created a Coffee object to continue with inventory tests.